+++ .github/workflows/install_mongo.sh
@@ -9,6 +9,8 @@ if [[ "$MONGODB" == *"4.2"* ]]; then
   mongo_build=mongodb-linux-x86_64-ubuntu1804-v${MONGODB}-latest
 elif [[ "$MONGODB" == *"4.4"* ]]; then
   mongo_build=mongodb-linux-x86_64-ubuntu1804-v${MONGODB}-latest
+elif [[ "$MONGODB" == *"5.0"* ]]; then
+  mongo_build=mongodb-linux-x86_64-ubuntu1804-v${MONGODB}-latest
 fi
 
 wget http://fastdl.mongodb.org/linux/$mongo_build.tgz