+++ mongoengine/base/fields.py
@@ -52,20 +52,20 @@ class BaseField:
         :param required: If the field is required. Whether it has to have a
             value or not. Defaults to False.
         :param default: (optional) The default value for this field if no value
-            has been set (or if the value has been unset).  It can be a
+            has been set, if the value is set to None or has been unset. It can be a
             callable.
-        :param unique: Is the field value unique or not.  Defaults to False.
+        :param unique: Is the field value unique or not (Creates an index).  Defaults to False.
         :param unique_with: (optional) The other field this field should be
-            unique with.
-        :param primary_key: Mark this field as the primary key. Defaults to False.
+            unique with (Creates an index).
+        :param primary_key: Mark this field as the primary key ((Creates an index)). Defaults to False.
         :param validation: (optional) A callable to validate the value of the
             field.  The callable takes the value as parameter and should raise
             a ValidationError if validation fails
         :param choices: (optional) The valid choices
-        :param null: (optional) If the field value can be null. If no and there is a default value
-            then the default value is set
+        :param null: (optional) If the field value can be null when a default exist. If not set, the default value
+        will be used in case a field with a default value is set to None. Defaults to False.
         :param sparse: (optional) `sparse=True` combined with `unique=True` and `required=False`
-            means that uniqueness won't be enforced for `None` values
+            means that uniqueness won't be enforced for `None` values (Creates an index). Defaults to False.
         :param **kwargs: (optional) Arbitrary indirection-free metadata for
             this field can be supplied as additional keyword arguments and
             accessed as attributes of the field. Must not conflict with any
@@ -282,6 +282,18 @@ class ComplexBaseField(BaseField):
         )
         return documents
 
+    def __set__(self, instance, value):
+        # Some fields e.g EnumField are converted upon __set__
+        # So it is fair to mimic the same behavior when using e.g ListField(EnumField)
+        EnumField = _import_class("EnumField")
+        if self.field and isinstance(self.field, EnumField):
+            if isinstance(value, (list, tuple)):
+                value = [self.field.to_python(sub_val) for sub_val in value]
+            elif isinstance(value, dict):
+                value = {key: self.field.to_python(sub) for key, sub in value.items()}
+
+        return super().__set__(instance, value)
+
     def __get__(self, instance, owner):
         """Descriptor to automatically dereference references."""
         if instance is None:
@@ -434,12 +446,12 @@ class ComplexBaseField(BaseField):
                             " have been saved to the database"
                         )
 
-                    # If its a document that is not inheritable it won't have
+                    # If it's a document that is not inheritable it won't have
                     # any _cls data so make it a generic reference allows
                     # us to dereference
                     meta = getattr(v, "_meta", {})
                     allow_inheritance = meta.get("allow_inheritance")
-                    if not allow_inheritance and not self.field:
+                    if not allow_inheritance:
                         value_dict[k] = GenericReferenceField().to_mongo(v)
                     else:
                         collection = v._get_collection_name()
@@ -509,14 +521,17 @@ class ObjectIdField(BaseField):
         return value
 
     def to_mongo(self, value):
-        if not isinstance(value, ObjectId):
-            try:
-                return ObjectId(str(value))
-            except Exception as e:
-                self.error(str(e))
-        return value
+        if isinstance(value, ObjectId):
+            return value
+
+        try:
+            return ObjectId(str(value))
+        except Exception as e:
+            self.error(str(e))
 
     def prepare_query_value(self, op, value):
+        if value is None:
+            return value
         return self.to_mongo(value)
 
     def validate(self, value):