+++ tests/fields/test_file_field.py
@@ -328,7 +328,7 @@ class TestFileField(MongoDBTestCase):
         assert len(list(files)) == 1
         assert len(list(chunks)) == 1
 
-        # Deleting the docoument should delete the files
+        # Deleting the document should delete the files
         testfile.delete()
 
         files = db.fs.files.find()