+++ tests/queryset/test_queryset.py
@@ -24,6 +24,7 @@ from mongoengine.queryset import (
     QuerySetManager,
     queryset_manager,
 )
+from mongoengine.queryset.base import BaseQuerySet
 from tests.utils import (
     requires_mongodb_gte_42,
     requires_mongodb_gte_44,
@@ -149,6 +150,14 @@ class TestQueryset(unittest.TestCase):
 
         persons = list(self.Person.objects().limit(0))
         assert len(persons) == 2 == n_docs
+
+    def test_limit_0(self):
+        """Ensure that QuerySet.limit works as expected."""
+        self.Person.objects.create(name="User A", age=20)
+
+        # Test limit with 0 as parameter
+        qs = self.Person.objects.limit(0)
+        assert qs.count() == 0
 
     def test_limit(self):
         """Ensure that QuerySet.limit works as expected."""
@@ -3505,6 +3514,27 @@ class TestQueryset(unittest.TestCase):
         foo.save()
 
         assert Foo.objects.distinct("bar") == [bar]
+        assert Foo.objects.no_dereference().distinct("bar") == [bar.pk]
+
+    def test_base_queryset_iter_raise_not_implemented(self):
+        class Tmp(Document):
+            pass
+
+        qs = BaseQuerySet(document=Tmp, collection=Tmp._get_collection())
+        with pytest.raises(NotImplementedError):
+            _ = list(qs)
+
+    def test_search_text_raise_if_called_2_times(self):
+        class News(Document):
+            title = StringField()
+            content = StringField()
+            is_active = BooleanField(default=True)
+
+        News.drop_collection()
+        with pytest.raises(OperationError):
+            News.objects.search_text("t1", language="portuguese").search_text(
+                "t2", language="french"
+            )
 
     def test_text_indexes(self):
         class News(Document):
@@ -3711,6 +3741,7 @@ class TestQueryset(unittest.TestCase):
         foo.save()
 
         assert Foo.objects.distinct("bar_lst") == [bar_1, bar_2]
+        assert Foo.objects.no_dereference().distinct("bar_lst") == [bar_1.pk, bar_2.pk]
 
     def test_custom_manager(self):
         """Ensure that custom QuerySetManager instances work as expected."""
@@ -3896,6 +3927,10 @@ class TestQueryset(unittest.TestCase):
         assert objects[post_1.id].title == post_1.title
         assert objects[post_2.id].title == post_2.title
         assert objects[post_5.id].title == post_5.title
+
+        objects = BlogPost.objects.as_pymongo().in_bulk(ids)
+        assert len(objects) == 3
+        assert isinstance(objects[post_1.id], dict)
 
         BlogPost.drop_collection()
 
@@ -4234,14 +4269,14 @@ class TestQueryset(unittest.TestCase):
         assert [1, 2, 3] == numbers
         Number.drop_collection()
 
-    def test_ensure_index(self):
+    def test_create_index(self):
         """Ensure that manual creation of indexes works."""
 
         class Comment(Document):
             message = StringField()
             meta = {"allow_inheritance": True}
 
-        Comment.ensure_index("message")
+        Comment.create_index("message")
 
         info = Comment.objects._collection.index_information()
         info = [